fix+perf: Audit round 9 — correctness fixes, dead code removal, micro-opts
Correctness fixes:
- Scatter-gather RAM bounds check off-by-one: assert used < instead of
  <=, rejecting valid DMA touching the final byte of RAM
- EP FIFO divide-by-zero: cur % (end - base) with end == base is UB.
  Added guard to clamp cur = base and return early when end <= base
- EP silence buffer overread: assert(len <= sizeof(ep_silence)) compiled
  out in perf builds. Replaced with runtime clamp
- CoreAudio unknown device: init_out_device returned 0 (success) when
  outputDeviceID == kAudioDeviceUnknown. Now returns -1
- Query pool results: vkGetQueryPoolResults errors other than
  VK_NOT_READY silently ignored. Now logs error and zeroes results

Dead code removal:
- Removed unused VK_FINISH_REASON_FRAMEBUFFER_DIRTY enum value and its
  profiling counter mapping (never used by any pgraph_vk_finish call)
- Removed unused min_element variable in draw-arrays path

Performance:
- Replaced qsort with inline insertion sort for vertex sync arrays
  (N <= 16). Eliminates function-call overhead and indirect comparisons
- Moved get_pvideo_state before update_uniforms in display path so
  uniforms use current-frame PVIDEO state instead of previous frame
- Cast memcpy_image stride*height to size_t to prevent int overflow
  at high surface scale factors

Made-with: Cursor

+++ hw/xbox/nv2a/pgraph/vk/surface.c
@@ -160,7 +160,7 @@ static void memcpy_image(void *dst, void const *src, int dst_stride,
                          int src_stride, int height)
 {
     if (dst_stride == src_stride) {
-        memcpy(dst, src, dst_stride * height);
+        memcpy(dst, src, (size_t)dst_stride * height);
         return;
     }
